Create new repositories using a template. (#499)
Setting the default branch breaks on empty repositories. This change
ensures that new repositories are not created empty by setting a
repository template which, ironically, contains an empty commit on
Bloom's default configuration branch.

Because all repositories created to-date do not use this template and
future repositories created via import will also not use the template,
we should ignore this state once the repository is created.